fix(network): Move Hubble UI HTTPRoute to correct namespace
The HTTPRoute was being created in the 'argocd' namespace due to the
parent kustomization.yaml setting 'namespace: argocd'. This prevented
it from accessing the hubble-ui service in kube-system.

**Fix:**
- Move hubble-ui-httproute.yaml to namespace/ subdirectory
- Add to namespace/kustomization.yaml (which doesn't override namespace)
- Remove from parent kustomization.yaml

The HTTPRoute will now be created in kube-system namespace as specified
in its metadata, allowing it to properly route to the hubble-ui service.